Use libteckit .so files from dependency
Rather than make and bundle .so files that work with all platforms.

This is meant to be an improvement, but a better solution will be to
modify this package or FieldWorks to depend on libteckit0 and find
the .so files where libteckit0 installs them. See also a note in
ReadMe_linux.txt regarding finding .so files on LD_LIBRARY_PATH.