Fix paths in nuget package
nuget expect libraries in certain paths. This change adjusts the
nuget package to use these paths. This makes it possible to add the
package to an sdk-style .csproj file. We also add a .props that defines
a few properties and items that can be used to copy the files.
Plus some housekeeping.