Update mono csproj files, code to build
- Update EncCnvtrs target framework. 4.7.1 is the latest framework in
mono5-sil.
- Set LangVersion to 7.3, which is the latest C# version supported in
mono5-sil.
- Set `using`s back to older syntax to still compile with mono5-sil.
- In WebBrowserAdaptor.cs, mono5-sil has trouble with the ternary:

  > error CS0173: Type of conditional expression cannot be determined
  > because there is no implicit conversion between
  > 'SilEncConverters40.WebBrowserEdge' and
  > 'SilEncConverters40.WebBrowserInstructions')

  So just changing to an `if`.

- Use older constructor syntax.
- WebView2
- Add WebBrowser .cs files
- Update nunit, nunittestadapter nuget packages.
- Move from using .nuget/packages.config on Linux to sharing
src/*/packages.config with the Windows build. So nuget restore from
all such packages.config files in Makefile.in. And add Geckofx Linux
targets. Having Linux and Windows Geckofx packages should just be
irrelevant on the wrong platform but we can look closer if we start
copying the wrong things at build or publishing.
- The mono5 compiler does not define `__MonoCS__`. So any lines
checking `#if !__MonoCS__` isn't going to cause an OS-specific
difference. However, most of these didn't seem to be a problem to
remove. The code in TestEncConverters.cs still needs some work to make
it work on Linux and so we still need to not run that code at present.
But I am changing how it gets disabled, to not use __MonoCS__.