build: Use same csproj files in Windows and Linux
- The Windows build and Linux builds have different .csproj files.
This requires extra maintenence to make changes to each, and pieces
can be missed at first. But the two platforms can share the same
csproj files, with some OS conditionals.
- Currently 'make' for the Linux build is using build/EC.proj, not
EncConverters-mono.sln, so this change carries on with using EC.proj.
- Two projects were not included on Linux, which are now added to
EC.proj: SilIndicEncConverters and DriveAiEncConverter. Though
DriveAiEncConverter is for testing and we could further adjust to not
include or ship it.
- rm *-mono.csproj files
- Change TargetFrameworkVersion down from 4.8 to 4.7.1 so it can work
with mono5-sil.
- Set LangVersion to 7.3 so it can work with mono5-sil.
- Generally, the Windows .csproj files were using a Platform of x86
and x64, and the Linux .csproj files were using a Platform of AnyCPU.
This made it easy to just add in a PropertyGroup for Debug|AnyCPU and
Release|AnyCPU that has the settings for the Linux build (as copied
from the foo-mono.csproj file into the 'foo 2010.csproj' file). One
important difference between the different Platform PropertyGroup
sections is the OutputPath, which is currently configured differently
for Windows and Linux. A further step will be to set OutputPath in a
way that works on both OSes. (Possibly with the help of a variable,
but perhaps just by having each OS output to
'..\..\output\$(Configuration)', unless it's important for other
elements to be included in that path.)
- Updated older 'System.XML' reference to 'System.Xml'.
- Where foo-mono.csproj specified CodeAnalysisModuleSuppressionsFile
or CodeAnalysisUseTypeNameInSuppression, I omitted those.

EC.proj:
- Add SilIndicEncConverters and DriveAiEncConverter.
- Switch from *-mono.csproj to '* 2010.csproj'.

'IcuEC 2010.csproj':
- Not compiling IcuBreakIterator*.cs or referencing ICU4NET.vcxproj on
Linux until that's fixed up.

'RunTests 2010.csproj':
- Windows was already building with Platform AnyCPU, so I added
sections for Linux and added to the Condition attributes whether the
OS was Linux or Windows.