build: handle OutputPath to share more of .csproj's between OSes
- From a recent change to make both Linux and Windows use the same
.csproj files, separate PropertyGroup sections were added with
Platform of AnyCPU for Linux. But it can be simpler if we handle the
OutputPath.
- Remove separate PropertyGroup sections for an AnyCPU configuration.
- Specify x64 as the Platform on Linux (Makefile.in). Note that Linux
FieldWorks is only shipped as 64b.
- The .csproj files used in Windows define OutputPath using
SolutionDir. One easy way to unify the OSes usage of the .csproj files
is to just set SolutionDir for Linux/mono. (Makefile.in).
- Update the output directory in other places on Linux from
output/Configuration to output/x64/Configuration (Makefile.in files,
EC.proj).
- RunTests.exe isn't build to the same output directory because we
don't need to ship it.